Added additional features
The map now reads individual county level data for the number of letters written. Added color with a legend to describe the number of letters written per state/county.

If we have time, I could add a smaller color legend to the counties which would make them more descriptive.

Note: You can zoom out of states and/or counties by clicking anywhere on the page past the edge of the country and surrounding states.